Create a base spider interface to rule them all
