+++ wsf_scraping/spiders/base_spider.py
@@ -1,0 +1,30 @@
+import scrapy
+from scrapy.spidermiddlewares.httperror import HttpError
+from twisted.internet.error import DNSLookupError
+from twisted.internet.error import TimeoutError
+
+
+class BaseSpider(scrapy.Spider):
+
+    def __init__(self, *args, **kwargs):
+        id = kwargs.get('uuid', '')
+        self.uuid = id
+
+    def on_error(self, failure):
+        self.logger.error(repr(failure))
+
+        if failure.check(HttpError):
+            response = failure.value.response
+            self.logger.error('HttpError on %s', response.url)
+
+        elif failure.check(DNSLookupError):
+            request = failure.request
+            self.logger.error('DNSLookupError on %s', request.url)
+
+        elif failure.check(TimeoutError):
+            request = failure.request
+            self.logger.error('TimeoutError on %s', request.url)
+
+    def _check_headers(self, response_headers):
+        content_type = response_headers.get('content-type', '').split(b';')[0]
+        return b'application/pdf' == content_type
